+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 **Features**
 
 - Added `/career/builder` page with Builder.io blog post snapshots and freshness disclosure. (#65)
-- Added Builder.io scraper tool to generate `content/career/builder.json` snapshots. (#65)
+- Added Builder.io scraper tool to generate `data-snapshots/career/builder-posts.json` snapshots. (#65)
 
 ## 28 Dec 2025